feat: added clone for specific remote branch

+++ tease_cli/src/commands/clone.rs
@@ -2,11 +2,23 @@ use std::{fs::{remove_dir_all, read_to_string}, path::Path, process::Command};
 
 use tease_common::write::bolb_writer::create_tease_file;
 
-use crate::remote_req::get_clone::get_clone;
+use crate::remote_req::{get_clone::get_clone, login::login_with_prompt};
+
+use super::create;
 
 pub fn clone(origin: String) {
+
+    let repo_name = origin.split("/").last().unwrap().to_string();
+    create::create_repo(repo_name.to_string());
+
+    let login_success = login_with_prompt(repo_name.to_string());
+    if !login_success.1 {
+        println!("Unauthorized.");
+        return ;
+    }
+
     let repo_name = origin.split("/").last().unwrap();
-    let res = get_clone(origin.to_string());
+    let res = get_clone(origin.to_string(), repo_name.to_string());
     if res.is_err() {
         println!("Couldn't clone {}", origin.to_string());
         let rem_res = remove_dir_all(repo_name.to_string());
@@ -28,14 +40,14 @@ pub fn clone(origin: String) {
     let maser_origin_path = format!("{}-origin", maser_path.to_string());
     create_tease_file(Path::new(&maser_origin_path.to_string()), head_commit.to_string());
 
-    execute_go_back_shell(repo_name.to_string(), head_commit.to_string());
+    execute_go_back_shell(repo_name.to_string(), head_commit.to_string(), login_success.0);
 }
 
-fn execute_go_back_shell(repo_name: String, head_commit: String) {
+fn execute_go_back_shell(repo_name: String, head_commit: String, email: String) {
     let go_back_res = Command::new("go_back".to_string())
         .arg(repo_name.to_string())    
         .arg(head_commit.to_string())
-        // .arg(head_commit.to_string())
+        .arg(email.to_string())
         .spawn();
     
     if go_back_res.is_err() {
